+++ src/test/unit/plugin_tests/table_test.js
@@ -859,8 +859,8 @@ test("Tab from th to next row", function () {
     expect(3);
     var expectedSelector = '#span_2_1';
     if (jQuery.browser.mozilla ||
-        (jQuery.browser.msie && parseInt(jQuery.browser.version, 10) >= 9.0))
-    {
+        (jQuery.browser.msie && parseInt(jQuery.browser.version, 10) >= 9.0)
+    ) {
         expectedSelector = '#td_2_1';
     }
     testTableTab(thTableHtml, '#th_1_3', expectedSelector);