+++ src/test/unit/specific_feature_tests/xml_parser.js
@@ -6,6 +6,19 @@ test, expect, deepEqual */
 
 module("XmlParser", {setup: setupWym});
 
+test("Empty is empty", function () {
+    var wymeditor = jQuery.wymeditors(0);
+    expect(2);
+
+    wymeditor._html('');
+    wymEqual(wymeditor, '');
+
+    wymeditor._html('');
+    // Placing the caret inside shouldn't create any content.
+    wymeditor.setCaretIn(jQuery(wymeditor._doc).find('body.wym_iframe')[0]);
+    wymEqual(wymeditor, '');
+});
+
 test("Should correct orphaned sublists", function () {
     expect(2);
     var expected = String() +