+++ src/test/unit/utils.js
@@ -332,14 +332,18 @@ function makeTextSelection(
 function moveSelector(wymeditor, selectedNode) {
     // This function was rewritten. Some of the existing callers were expecting
     // assertions and others were not. Next line handles this gracefully.
-    if (expect()) {expect(expect() - 1);}
+    if (expect()) {
+        expect(expect() - 1);
+    }
 
     if (
         wymeditor.canSetCaretIn(selectedNode)
     ) {
         wymeditor.setCaretIn(selectedNode);
 
-        if (expect()) {expect(expect() + 1);}
+        if (expect()) {
+            expect(expect() + 1);
+        }
 
         deepEqual(
             wymeditor.selectedContainer(),
@@ -352,7 +356,9 @@ function moveSelector(wymeditor, selectedNode) {
 
         wymeditor.setCaretBefore(selectedNode);
 
-        if (expect()) {expect(expect() + 1);}
+        if (expect()) {
+            expect(expect() + 1);
+        }
 
         deepEqual(
             wymeditor.nodeAfterSel(),