+++ src/wymeditor/editor/base.js
@@ -203,16 +203,38 @@ WYMeditor.editor.prototype.init = function () {
     // to work-around old ones and predict new ones, let's just ensure the
     // initialization only happens once. All methods of detecting load are
     // unreliable.
-    wym._iframe_initialized = false;
+    wym.iframeInitialized = false;
 
     jQuery(wym._box).find('iframe').load(function () {
-        if (wym._iframe_initialized === true) {
+        if (wym.iframeInitialized === true) {
             return;
         }
-        wym._iframe_initialized = wym.initIframe(this);
+        wym.initIframe(this);
     });
 
     wym.initSkin();
+};
+
+/**
+    WYMeditor.editor.postIframeInit
+
+    Part of the editor's initialization, which must be done after the
+    iframe's initialization.
+*/
+WYMeditor.editor.prototype.postIframeInit = function () {
+    var wym = this;
+
+    if (jQuery.isFunction(wym._options.postInit)) {
+        wym._options.postInit(wym);
+    }
+
+    // Add event listeners to doc elements, e.g. images
+    wym.listen();
+
+    jQuery(wym._element).trigger(
+        WYMeditor.EVENTS.postIframeInitialization,
+        wym._wym
+    );
 };
 
 /**
@@ -478,22 +500,34 @@ WYMeditor.editor.prototype.nodeAfterSel = function () {
 WYMeditor.editor.prototype.selectedContainer = function () {
     var focusNode = this.selection().focusNode;
 
-        if (
-            focusNode.nodeType === WYMeditor.NODE.TEXT || (
+    if (
+        focusNode.nodeType === WYMeditor.NODE.TEXT || (
 
-                focusNode.tagName &&
+            focusNode.tagName &&
 
-                jQuery.inArray(
-                    focusNode.tagName.toLowerCase(),
-                    WYMeditor.NON_CONTAINING_ELEMENTS
-                ) > -1
-            )
-        ) {
-            return focusNode.parentNode;
-        } else {
-            return focusNode;
-        }
+            jQuery.inArray(
+                focusNode.tagName.toLowerCase(),
+                WYMeditor.NON_CONTAINING_ELEMENTS
+            ) > -1
+        )
+    ) {
+        return focusNode.parentNode;
+    } else {
+        return focusNode;
+    }
 
+};
+
+// Deprecated in favor of `WYMeditor.editor.selectedContainer`.
+WYMeditor.editor.prototype.selected = function () {
+    var wym = this;
+
+    WYMeditor.console.warn(
+        "The function WYMeditor.editor.selected() is " +
+        "deprecated. Use WYMeditor.editor.selectedContainer"
+    );
+
+    return wym.selectedContainer();
 };
 
 /**
@@ -604,7 +638,7 @@ WYMeditor.editor.prototype.isInlineNode = function (node) {
 
 /**
     WYMeditor.editor.isListNode
-    =============================
+    ===========================
 
     Returns true if the provided node is a list element. Otherwise
     returns false.
@@ -612,7 +646,7 @@ WYMeditor.editor.prototype.isInlineNode = function (node) {
     @param node The node to check.
 */
 
-WYMeditor.editor.prototype.isListNode= function (node) {
+WYMeditor.editor.prototype.isListNode = function (node) {
     if (
         node.tagName &&
         jQuery.inArray(
@@ -623,6 +657,93 @@ WYMeditor.editor.prototype.isListNode= function (node) {
         return true;
     }
     return false;
+};
+
+/**
+   WYMeditor.editor.unwrapIfMeaninglessSpan
+
+   If the given node is a span with no useful attributes, unwrap it.
+
+   For certain editing actions (mostly list indent/outdent), it's necessary to
+   wrap content in a span element to retain grouping because it's not obvious that
+   the content will stay together without grouping. This method detects that
+   specific situation and then unwraps the content if the span is in fact not
+   necessary. It handles the fact that IE7 throws attributes on spans, even if
+   they're completely empty.
+
+   Unlike sane browsers, IE7 provides all possible attributes in a node's
+   `attributes`. Thus, a simple check like `attributes.length > 0` isn't
+   enough (as long as we support IE7). This function tries to determine
+   whether the element really does have meaningful attributes, considering
+   IE7's behavior. This turns out to be not so simple, as IE7 populates
+   some attributes with truth-y values. So the mechanism is a kind of
+   a blacklist filter of IE7's junk-attributes. If an attribute passes this
+   blacklist filter, then this function returns true.
+
+   @param element The element.
+*/
+WYMeditor.editor.prototype.unwrapIfMeaninglessSpan = function (element) {
+    var $element = jQuery(element),
+        attributes,
+        i,
+        attrName,
+        attrValue,
+        // We don't care about any of these attributes
+        meaninglessAttrNames = [
+            '_wym_visited',
+            'dataFld',
+            'onmouseup',
+            'contentEditable',
+            'dataFormatAs',
+            'dataSrc',
+            'tabIndex',
+            'value'
+        ],
+        // Any attribute with these values isn't interesting
+        falsyAttrValues = [
+            '',
+            undefined,
+            false,
+            null
+        ];
+
+    if (!element || typeof (element.tagName) === 'undefined' ||
+        element.tagName.toLowerCase() !== 'span') {
+        return false;
+    }
+
+    attributes = element.attributes;
+    if (attributes.length === 0) {
+        // Early return for spans with no attributes
+        $element.before($element.contents());
+        $element.remove();
+        return true;
+    }
+
+    // This loop is required for IE7 because it seems to populate
+    // the attributes property with all possible attributes. When
+    // support for IE7 is dropped the length check should be
+    // enough.
+    for (i = 0; i < attributes.length; i++) {
+        attrName = attributes[i].name;
+        // Getting the value through jQuery is rumored to normalizes it in some
+        // ways.
+        attrValue = $element.attr(attrName);
+        if (
+            jQuery.inArray(attrName, meaninglessAttrNames) === -1 &&
+            jQuery.inArray(
+                attrValue,
+                falsyAttrValues
+            ) === -1
+        ) {
+            // We hit an attribute making this non-meaningless
+            return false;
+        }
+    }
+
+    $element.before($element.contents());
+    $element.remove();
+    return true;
 };
 
 /**
@@ -725,6 +846,17 @@ WYMeditor.editor.prototype.mainContainer = function (sType) {
     }
 
     return false;
+};
+
+// Deprecated in favor of `WYMeditor.editor.mainContainer`.
+WYMeditor.editor.prototype.container = function (sType) {
+    var wym = this;
+
+    WYMeditor.console.warn(
+        "The function `WYMeditor.editor.container` is " +
+        "deprecated. Use `WYMeditor.editor.mainContainer`.");
+
+    return wym.mainContainer(sType);
 };
 
 /**
@@ -837,31 +969,55 @@ WYMeditor.editor.prototype.findUp = function (node, filter) {
     WYMeditor.editor.switchTo
     =========================
 
-    Switch the type of the given `node` to type `sType`. If stripAttrs is true,
-    the attributes of node will not be included in the the new type. If
-    stripAttrs is false (or undefined), the attributes of node will be
-    preserved through the switch.
-*/
-WYMeditor.editor.prototype.switchTo = function (node, sType, stripAttrs) {
-    var newNode = this._doc.createElement(sType),
-        html = jQuery(node).html(),
-        attrs = node.attributes,
-        i;
+    Switch the type of an element.
 
-    if (node.tagName.toLowerCase() === 'img') {
-        throw "Will not change the tag of this element.";
+    @param element The element.
+    @param sType A string of the desired type. For example, 'p'.
+    @param stripAttrs a boolean that determines whether the attributes of
+                      the element will be stripped or preserved.
+    @param setCaret A boolean that determines whether the caret will be set at
+                    the beginning, inside the element, after the switch. Default
+                    is false.
+
+*/
+WYMeditor.editor.prototype.switchTo = function (
+    element,
+    sType,
+    stripAttrs,
+    setCaret
+) {
+    var wym = this,
+        $element = jQuery(element),
+        newElement,
+        i,
+        attrs = element.attributes;
+
+    if (!element.tagName) {
+        throw "This must be an element.";
     }
+
+    if (element.tagName.toLowerCase() === 'img') {
+        throw "Will not change the type of an 'img' element.";
+    }
+
+    newElement = wym._doc.createElement(sType);
+    jQuery(newElement).append(element.childNodes);
+    $element.replaceWith(newElement);
 
     if (!stripAttrs) {
         for (i = 0; i < attrs.length; ++i) {
-            newNode.setAttribute(attrs.item(i).nodeName,
-                                 attrs.item(i).nodeValue);
+            newElement.setAttribute(
+                attrs.item(i).nodeName,
+                attrs.item(i).nodeValue
+            );
         }
     }
-    newNode.innerHTML = html;
-    node.parentNode.replaceChild(newNode, node);
 
-    this.setCaretIn(newNode);
+    if (setCaret) {
+        wym.setCaretIn(newElement);
+    }
+
+    return newElement;
 };
 
 WYMeditor.editor.prototype.replaceStrings = function (sVal) {
@@ -2004,20 +2160,20 @@ WYMeditor.editor.prototype._isPOrDivAfterEnterInEmptynestedLi = function
         container.parentNode.tagName.toLowerCase() === 'li'
     ) {
         switch (container.childNodes.length) {
-            case 0:
+        case 0:
+            return true;
+        case 1:
+            if (
+                container.childNodes[0].tagName &&
+                container.childNodes[0].tagName.toLowerCase() === 'br'
+            ) {
                 return true;
-            case 1:
-                if (
-                    container.childNodes[0].tagName &&
-                    container.childNodes[0].tagName.toLowerCase() === 'br'
-                ) {
-                    return true;
-                } else if (
-                    container.childNodes[0].nodeType === WYMeditor.NODE.TEXT &&
-                    container.childNodes[0].data === WYMeditor.NBSP
-                ) {
-                    return true;
-                }
+            } else if (
+                container.childNodes[0].nodeType === WYMeditor.NODE.TEXT &&
+                container.childNodes[0].data === WYMeditor.NBSP
+            ) {
+                return true;
+            }
         }
     }
     return false;
@@ -2038,7 +2194,7 @@ WYMeditor.editor.prototype._isPOrDivAfterEnterInEmptynestedLi = function
  */
 
 WYMeditor.editor.prototype._isSpilledListAfterEnterInEmptyLi = function
-(container) {
+    (container) {
     if (
         container.tagName.toLowerCase() === 'li' &&
         container.previousSibling &&
@@ -2113,10 +2269,10 @@ WYMeditor.editor.prototype._replaceNodeWithBrAndSetCaret = function (node) {
     ) {
         $node.before('<br />');
     }
-        $node.before('<br />');
-        this.setCaretBefore(node.previousSibling);
-        $node.remove();
 
+    $node.before('<br />');
+    this.setCaretBefore(node.previousSibling);
+    $node.remove();
 };
 
 /**
@@ -2410,119 +2566,75 @@ WYMeditor.editor.prototype.getCommonParentList = function (listItems, getClosest
     editor._getSelectedListItems
     ============================
 
-    Based on the given selection, determine which li nodes are "selected" from
-    the user's standpoint. These are the li nodes that they would expect to be
-    affected by an action with the given selection.
+    Determine which `li` nodes are "selected" from the user's standpoint.
 
-    Generally, this means any li which has at least some of its text content
-    highlighted will be returned.
+    These are the `li` nodes that they would expect to be affected by an action
+    with the given selection.
+
+    For a better understanding, comments are provided inside.
+
+    @param selection A Rangy Selection object.
 */
-WYMeditor.editor.prototype._getSelectedListItems = function (sel) {
+WYMeditor.editor.prototype._getSelectedListItems = function (selection) {
     var wym = this,
-        i,
-        j,
-        range,
-        selectedLi,
-        nodes = [],
-        liNodes = [],
-        containsNodeTextFilter,
-        parentsToAdd,
-        node;
+        $selectedContainer,
+        $selectedNodes,
+        $selectedLis;
 
-    // Filter function to remove undesired nodes from what rangy.getNodes()
-    // gives
-    containsNodeTextFilter = function (testNode) {
-        var fullyContainsNodeText;
+    if (selection.isCollapsed) {
+        $selectedContainer = jQuery(wym.selectedContainer());
 
-        // Include any partially-selected textNodes
-        if (rangy.dom.isCharacterDataNode(testNode)) {
-            return testNode;
+        if ($selectedContainer.closest('li, table').is('table')) {
+            // Inside a table and not inside a list inside it. This prevents
+            // the inclusion of the list item that might be an ancestor of the
+            // table.
+            return [];
         }
-
-        try {
-            fullyContainsNodeText = range.containsNodeText(testNode);
-        } catch (e) {
-            // Rangy throws an exception on an internal
-            // intersection call on the last node that's
-            // actually in the selection
-            return true;
-        }
-
-        if (fullyContainsNodeText === true) {
-            // If we fully contain any text in this node, it's definitely
-            // selected
-            return true;
-        }
-    };
-
-    // Iterate through all of the selection ranges and include any li nodes
-    // that are user-selected in each range
-    for (i = 0; i < sel.rangeCount; i++) {
-        range = sel.getRangeAt(i);
-        if (range.collapsed === true) {
-            // Collapsed ranges don't return the range they're in as part of
-            // getNodes, so let's find the next list item up
-            selectedLi = wym.findUp(range.startContainer, 'li');
-            if (selectedLi) {
-                nodes = nodes.concat([selectedLi]);
-            }
-        } else {
-            // getNodes includes the parent list item whenever we have our
-            // selection in a sublist. We need to make a distinction between
-            // when the parent list item is actually selected and when it's
-            // only sort of selected because we're selecting a sub-item
-            // (meaning it's partially selected).
-            // In the following case, we don't want `2` as one of our nodes:
-            // 1
-            // 2
-            //   2.1
-            //   2|.2
-            // 3|
-            nodes = nodes.concat(
-                range.getNodes(
-                    [],
-                    containsNodeTextFilter
-                )
-            );
-
-            // We also need to include the parent li if we selected a non-li, non-list node.
-            // eg. if we select text inside an li, the user is actually
-            // selecting that entire li
-            parentsToAdd = [];
-            for (j = 0; j < nodes.length; j++) {
-                node = nodes[j];
-                if (!jQuery(node).is('li,ol,ul')) {
-                    // Crawl up the dom until we find an li
-                    while (node.parentNode) {
-                        node = node.parentNode;
-                        if (jQuery(node).is('li')) {
-                            parentsToAdd.push(node);
-                            break;
-                        }
-                    }
-                }
-            }
-            // Add in all of the new parents if they're not already included
-            // (no duplicates)
-            for (j = 0; j < parentsToAdd.length; j++) {
-                if (jQuery.inArray(parentsToAdd[j], nodes) === -1) {
-                    nodes.push(parentsToAdd[j]);
-                }
-            }
-
-
-        }
+        return $selectedContainer.closest('li');
     }
 
-    // Filter out the non-li nodes
-    for (i = 0; i < nodes.length; i++) {
-        if (nodes[i].nodeType === WYMeditor.NODE.ELEMENT &&
-                nodes[i].tagName.toLowerCase() === WYMeditor.LI) {
-            liNodes.push(nodes[i]);
-        }
+    // All the selected nodes in the selection's first range.
+    $selectedNodes = jQuery(selection.getRangeAt(0).getNodes());
+
+    if ($selectedNodes.closest('li, table').filter('li').length === 0) {
+        // Selection is in a table before it is in a list. This prevents
+        // inclusion of the list item that the table may be contained
+        // in.
+        return [];
     }
 
-    return liNodes;
+    // The technique is to get selected contents of the list items and then
+    // get their closest parent list items. So we don't want the list elements.
+    // Some list items may be empty and we do want those so we'll add them back
+    // later.
+    $selectedLis = $selectedNodes.not('li, ol, ul')
+
+    // IE doesn't include the Rangy selection boundary `span` in the above
+    // `.getNodes`. This effects an edge case, where a `li` contains only
+    // that `span`.
+    .add($selectedNodes.find('.rangySelectionBoundary'))
+
+    // Add back the text nodes because jQuery.not always excludes them.
+    .add(
+        $selectedNodes.filter(
+            function () {
+                return wym.nodeType === WYMeditor.NODE.TEXT;
+            }
+        )
+    )
+    .closest('li')
+
+    // Add `li`s that are selected and are empty. Because they didn't
+    // get found by `.closest`. We can safely add them because they don't
+    // contain other list items so surely if they are in the selection it is
+    // because the user wants to manipulate them.
+    .add($selectedNodes.filter('li:empty'))
+
+    // Exclude list items that are children of table elements that are in the
+    // selection.
+    .not($selectedNodes.filter('table').find('li'));
+
+    return jQuery.unique($selectedLis.get());
 };
 
 /**
@@ -2690,7 +2802,7 @@ WYMeditor.editor.prototype.outdent = function () {
 
     A helper function to ensure that the selection is restored to the same
     location after a potentially-complicated dom manipulation is performed. This
-    also handles the case where the dom manipulation throws and error by cleaning
+    also handles the case where the dom manipulation throws an error by cleaning
     up any selection markers that were added to the dom.
 
     `manipulationFunc` is a function that takes no arguments and performs the
@@ -2803,53 +2915,79 @@ WYMeditor.editor.prototype.insertUnorderedlist = function () {
     editor._insertList
     ==================
 
-    Convert the selected block in to the specified type of list.
+    This either manipulates existing lists or creates a new one.
 
-    If the selection is already inside a list, switch the type of the nearest
-    parent list to an `<ol>`. If the selection is in a block element that can be a
-    valid list, place that block element's contents inside an ordered list.
+    The action that will be performed depends on the contents of the
+    selection and their context.
+
+    This can result in one of:
+
+     1. Changing the type of lists.
+     2. Removing items from list.
+     3. Creating a list.
+     4. Nothing.
+
+    If existing list items are selected this means either changing list type
+    or de-listing. Changing list type occurs when selected list items all share
+    a list of a different type than the requested. Removing items from lists
+    occurs when selected list items are all of the same type as the requested.
+
+    If no list items are selected, then, if possible, a list will be created.
+    If not possible, no change is made.
 
     Returns `true` if a change was made, `false` otherwise.
+
+    @param listType A string, representing the user's action, either 'ul'
+                    or 'ol'.
  */
 WYMeditor.editor.prototype._insertList = function (listType) {
     var wym = this._wym,
-        sel = rangy.getIframeSelection(this._iframe),
+        sel = rangy.getIframeSelection(wym._iframe),
         listItems,
-        rootList,
-        selectedBlock,
+        $listItems,
+        $parentListsOfDifferentType,
+        commonParentList,
         potentialListBlock;
 
     listItems = wym._getSelectedListItems(sel);
+    if (listItems.length > 0) {
+        // We have existing list items selected. This means either changing
+        // their parent lists' types or de-listing.
 
-    // If we've selected some list items all in the same list, we want to
-    // change the type of that list.
-    if (listItems.length !== 0) {
-        // If the selection is across paragraphs and other items at the root level,
-        // don't indent
-        rootList = wym.getCommonParentList(listItems, true);
-        if (rootList) {
-            this._changeListType(rootList, listType);
-            return true;
+        $listItems = jQuery(listItems);
+
+        $parentListsOfDifferentType = $listItems
+            .parent(':not(' + listType + ')');
+        if ($parentListsOfDifferentType.length > 0) {
+            // Some of the lists are of a different type than that which was
+            // requested.
+
+            // Change list type only if selected list items share a common parent
+            // list. TODO: Change types over several lists:
+            // https://github.com/wymeditor/wymeditor/issues/541
+            commonParentList = wym.getCommonParentList(listItems, true);
+            if (commonParentList) {
+                wym._changeListType(commonParentList, listType);
+                return true;
+            }
         } else {
-            // We have a selection across multiple root-level lists. Punt on
-            // this case for now.
-            // TODO: Handle multiple root-level lists properly
-            return false;
+            // List types are the same as requested. De-list.
+            wym._removeItemsFromList($listItems);
+            return true;
         }
-
     }
-
-    // If we've selected a block-level item that's appropriate to convert in to a list,
-    // convert it.
-    selectedBlock = this.selectedContainer();
+    // Get a potential block element from selection that could be converted
+    // into a list:
     // TODO: Use `_containerRules['root']` minus the ol/ul and
     // `_containerRules['contentsCanConvertToList']
-    potentialListBlock = this.findUp(selectedBlock, WYMeditor.POTENTIAL_LIST_ELEMENTS);
+    potentialListBlock = wym.findUp(
+        wym.selectedContainer(),
+        WYMeditor.POTENTIAL_LIST_ELEMENTS
+    );
     if (potentialListBlock) {
-        this._convertToList(potentialListBlock, listType);
+        wym._convertToList(potentialListBlock, listType);
         return true;
     }
-
     // The user has something selected that wouldn't be a valid list
     return false;
 };
@@ -2887,6 +3025,200 @@ WYMeditor.editor.prototype._convertToList = function (blockElement, listType) {
     $newList = $blockElement.children();
 
     return $newList.get(0);
+};
+
+/**
+    editor._removeItemsFromList
+    ===========================
+
+    De-list the provided list items.
+
+    @param listItems A jQuery object of list items.
+*/
+WYMeditor.editor.prototype._removeItemsFromList = function ($listItems) {
+    var wym = this,
+        $listItem,
+        i,
+        j,
+        listItemChild,
+        $childNodesToTransfer,
+        k,
+        $childNodeToTransfer,
+        rootContainer = wym.documentStructureManager.structureRules
+            .defaultRootContainer,
+        attributes;
+
+    // This is left here for future reference because it may or may not be a
+    // better behavior:
+    // It is reasonable to de-list only a subset of the provided list items.
+    // The subset are the list items which are not ancestors of any other list
+    // items.
+    //$listItems = $listItems.not($listItems.find('li'));
+
+    for (i = 0; i < $listItems.length; i++) {
+        $listItem = $listItems.eq(i);
+
+        // Determine the type of element this list item will be transformed
+        // into and call for this transformation.
+        if ($listItem.parent().parent('li, th, td').length === 1) {
+            // The list item will end up inside another list item or inside a
+            // table cell. Turn it into a `span`.
+            $listItem = jQuery(
+                wym.switchTo(
+                    $listItem[0],
+                    'span',
+                    false
+                )
+            );
+        } else {
+            // The list item will end up in the root of the document. Turn it
+            // into a default root container.
+            $listItem = jQuery(
+                wym.switchTo(
+                    $listItem[0],
+                    rootContainer,
+                    false
+                )
+            );
+        }
+        // The transformation should be complete and the element is no longer
+        // a list item, hence we call it 'the de-listed element'.
+
+        // Move the de-listed element according to its relation to its
+        // sibling nodes.
+        if ($listItem.parent().children().length === 1) {
+            // It is the only child in the list.
+
+            $listItem.parent().before($listItem);
+            // Remove the list because it is empty.
+            $listItem.next().remove();
+
+        } else if ($listItem[0] === $listItem.parent().children().first()[0]) {
+            // It is not the only child. It is the first child.
+
+            $listItem.parent().before($listItem);
+        } else if (
+            $listItem[0] !== $listItem.parent().children().first()[0] &&
+            $listItem[0] !== $listItem.parent().children().last()[0]
+        ) {
+            // It is not the first and not the last child.
+            $listItem.parent().before(
+                '<' + $listItem.parent()[0].tagName + '/>'
+            );
+            jQuery($listItem.prevAll().toArray().reverse())
+                .appendTo($listItem.parent().prev());
+            $listItem.parent().before($listItem);
+
+        } else if ($listItem[0] === $listItem.parent().children().last()[0]) {
+            // It is not the only child. It is the last child.
+            $listItem.parent().after($listItem);
+        }
+        // The de-listed element should now be at it's final destination.
+        // Now, deal with its contents.
+        for (j = 0; j < $listItem.contents().length; j++) {
+            listItemChild = $listItem.contents()[j];
+            if (
+                wym.isBlockNode(listItemChild) &&
+                // Prevents a rangy selection boundary element from interfering.
+                listItemChild.className !== 'rangySelectionBoundary' &&
+                listItemChild.tagName.toLowerCase() !== 'br'
+            ) {
+                // We have hit the first block child. From this child onward,
+                // the contents will be moved to after the de-listed element.
+                $childNodesToTransfer = $listItem.contents().slice(j);
+                if (
+                    $listItem[0].tagName.toLowerCase() ===
+                        rootContainer
+                ) {
+                    // The destination of these nodes is the root element.
+                    // Prepare them as such.
+                    for (k = 0; k < $childNodesToTransfer.length; k++) {
+                        $childNodeToTransfer = $childNodesToTransfer.eq(k);
+                        if (
+                            $childNodeToTransfer[0]
+                                .nodeType === WYMeditor.NODE.TEXT
+                        ) {
+                            $childNodeToTransfer.wrap(
+                                '<' + rootContainer + ' />'
+                            );
+                        } else if (
+                            $childNodeToTransfer[0].tagName &&
+                            !(wym.isBlockNode($childNodeToTransfer[0])) &&
+                            $childNodeToTransfer[0].tagName
+                                .toLowerCase() !== 'br'
+                        ) {
+                            wym.switchTo(
+                                $childNodesToTransfer[k],
+                                rootContainer,
+                                false
+                            );
+                        }
+                    }
+                }
+                // The contents should be ready now.
+                $listItem.after($listItem.contents().slice(j));
+                // The loop was for finding the first block element.
+                break;
+            }
+        }
+        // `br`s may have been transferred to the root container. They don't
+        // belong there.
+        jQuery(wym._doc).find('body.wym_iframe').children('br').remove();
+
+        if ($listItem[0].tagName.toLowerCase() === 'span') {
+            // Get rid of empty `span`s and ones that contain only `br`s.
+            if (
+                $listItem.contents(':not(.rangySelectionBoundary)')
+                    .length === 0 ||
+                $listItem.contents(':not(.rangySelectionBoundary)').length ===
+                    $listItem.contents('br').length
+            ) {
+                // The Rangy selection boundary `span` may be inside.
+                $listItem.before($listItem.contents('.rangySelectionBoundary'));
+                $listItem.remove();
+            } else {
+                // The `span` wasn't removed.
+
+                // Add `br` elements that may be necessary because by turning
+                // a `li` element into a `span` element we turn a block
+                // type element into an in-line type element.
+                if (
+                    $listItem[0].previousSibling &&
+                    $listItem[0].previousSibling.nodeType === WYMeditor
+                        .NODE.TEXT ||
+
+                    $listItem.prevAll(':not(.rangySelectionBoundary)')
+                        .length > 0 &&
+                    wym.isBlockNode(
+                        $listItem.prevAll(':not(.rangySelectionBoundary)')[0]
+                    ) === false
+                ) {
+                    $listItem.before('<br />');
+                }
+                if (
+                    $listItem[0].nextSibling &&
+                    $listItem[0].nextSibling.nodeType === WYMeditor
+                        .NODE.TEXT ||
+
+                    $listItem.nextAll(':not(.rangySelectionBoundary)')
+                        .length > 0 &&
+                    wym.isBlockNode(
+                        $listItem.nextAll(':not(.rangySelectionBoundary)')[0]
+                    ) === false
+                ) {
+                    $listItem.after('<br />');
+                }
+
+                // If the de-listed element has no meaningful attributes, there is
+                // no use for it being a span.
+                attributes = $listItem[0].attributes;
+                wym.unwrapIfMeaninglessSpan($listItem[0]);
+            }
+        }
+    }
+
+    // Reintroduce any necessary DOM-level corrections for editing purposes
+    wym.fixBodyHtml();
 };
 
 /**